+++ packages/yourloops/lib/patient/patient.hook.tsx
@@ -25,7 +25,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import React, { useCallback, useEffect, useMemo, useState } from 'react'
+import { useCallback, useEffect, useMemo, useState } from 'react'
 import moment from 'moment-timezone'
 
 import { type Team, useTeam } from '../team'
@@ -38,43 +38,44 @@ import metrics from '../metrics'
 import { errorTextFromException } from '../utils'
 import { type PatientContextResult } from './models/patient-context-result.model'
 import { type Patient } from './models/patient.model'
-import { type PatientListFilters } from '../../components/patient-list/enums/patient-list.enum'
+import { PatientListFilters } from '../../components/patient-list/enums/patient-list.enum'
 import { UserInvitationStatus } from '../team/models/enums/user-invitation-status.enum'
 import { type MedicalData } from '../data/models/medical-data.model'
 import { type PatientTeam } from './models/patient-team.model'
 import { useSelectedTeamContext } from '../selected-team/selected-team.provider'
 import { PRIVATE_TEAM_ID } from '../team/team.hook'
+import { type PatientFilterStats } from './models/patient-filter-stats.model'
+import { useAlert } from '../../components/utils/snackbar'
 
 export default function usePatientProviderCustomHook(): PatientContextResult {
   const { cancel: cancelInvitation, getInvitation, refreshSentInvitations } = useNotification()
   const { refresh: refreshTeams } = useTeam()
   const { user, getFlagPatients, flagPatient } = useAuth()
   const { selectedTeam } = useSelectedTeamContext()
+  const alert = useAlert()
 
   const [patients, setPatients] = useState<Patient[]>([])
   const [initialized, setInitialized] = useState<boolean>(false)
   const [refreshInProgress, setRefreshInProgress] = useState<boolean>(false)
-  const [errorMessage, setErrorMessage] = React.useState<string | null>(null)
 
   const fetchPatients = useCallback(() => {
     PatientUtils.computePatients(user).then(computedPatients => {
       setPatients(computedPatients)
-      setErrorMessage(null)
     }).catch((reason: unknown) => {
       const message = errorTextFromException(reason)
-      if (message !== errorMessage) {
-        setErrorMessage(message)
-      }
+      alert.error(message)
     }).finally(() => {
       setInitialized(true)
       setRefreshInProgress(false)
     })
-  }, [errorMessage, user])
+    // Need to rewrite the alert component, or it triggers infinite loop...
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [user])
 
-  const refresh = useCallback(() => {
+  const refresh = (): void => {
     setRefreshInProgress(true)
     fetchPatients()
-  }, [fetchPatients])
+  }
 
   const updatePatient = useCallback((patient: Patient) => {
     const patientsUpdated = patients.filter(p => p.userid !== patient.userid)
@@ -89,10 +90,8 @@ export default function usePatientProviderCustomHook(): PatientContextResult {
     return patientTeam.teamId === PRIVATE_TEAM_ID
   }
 
-  const buildPatientFiltersStats = useCallback(() => {
+  const buildPatientFiltersStats = useCallback((): PatientFilterStats => {
     return {
-      all: patientList.filter((patient) => !PatientUtils.isOnlyPendingInvitation(patient)).length,
-      pending: patientList.filter((patient) => PatientUtils.isInvitationPending(patient)).length,
       directShare: patientList.filter((patient) => patient.teams.find(team => isPatientTeamPrivate(team))).length,
       unread: patientList.filter(patient => patient.metadata.hasSentUnreadMessages).length,
       outOfRange: patientList.filter(patient => patient.alarms.timeSpentAwayFromTargetActive).length,
@@ -103,14 +102,16 @@ export default function usePatientProviderCustomHook(): PatientContextResult {
     }
   }, [patientList])
 
-  const patientsFilterStats = useMemo(() => buildPatientFiltersStats(), [buildPatientFiltersStats])
+  const patientsFilterStats = useMemo<PatientFilterStats>(() => buildPatientFiltersStats(), [buildPatientFiltersStats])
 
-  const getPatientByEmail = useCallback((email: string) => patientList.find(patient => patient.profile.email === email), [patientList])
+  const getPatientByEmail = (email: string): Patient => patientList.find(patient => patient.profile.email === email)
 
-  const getPatientById = useCallback(userId => patientList.find(patient => patient.userid === userId), [patientList])
+  const getPatientById = (userId: string): Patient => patientList.find(patient => patient.userid === userId)
 
-  const filterPatients = useCallback((filterType: PatientListFilters, search: string, flaggedPatients: string[]) => {
-    const filteredPatients = PatientUtils.extractPatients(patientList, filterType, flaggedPatients)
+  const pendingPatientsCount = selectedTeam ? PatientUtils.extractPatients(patientList, PatientListFilters.Pending, [], selectedTeam.id).length : undefined
+
+  const filterPatients = (filterType: PatientListFilters, search: string, flaggedPatients: string[]): Patient[] => {
+    const filteredPatients = PatientUtils.extractPatients(patientList, filterType, flaggedPatients, user.isUserHcp() ? selectedTeam.id : undefined)
 
     if (search.length === 0) {
       return filteredPatients
@@ -130,29 +131,29 @@ export default function usePatientProviderCustomHook(): PatientContextResult {
       const lastName = patient.profile.lastName ?? ''
       return firstName.toLocaleLowerCase().includes(searchText) || lastName.toLocaleLowerCase().includes(searchText)
     })
-  }, [patientList])
+  }
 
-  const invitePatient = useCallback(async (team: Team, username: string) => {
+  const invitePatient = async (team: Team, username: string): Promise<void> => {
     await PatientApi.invitePatient({ teamId: team.id, email: username })
     await refreshSentInvitations()
     refresh()
-  }, [refresh, refreshSentInvitations])
+  }
 
-  const editPatientRemoteMonitoring = useCallback((patient: Patient) => {
+  const editPatientRemoteMonitoring = (patient: Patient): void => {
     const patientUpdated = getPatientById(patient.userid)
     if (!patientUpdated) {
       throw Error(`Cannot update monitoring of patient with id ${patient.userid} as patient was not found`)
     }
     patientUpdated.monitoring = patient.monitoring
     updatePatient(patientUpdated)
-  }, [getPatientById, updatePatient])
+  }
 
   const markPatientMessagesAsRead = useCallback((patient: Patient) => {
     patient.metadata.hasSentUnreadMessages = false
     updatePatient(patient)
   }, [updatePatient])
 
-  const updatePatientMonitoring = useCallback(async (patient: Patient) => {
+  const updatePatientMonitoring = async (patient: Patient): Promise<void> => {
     if (!patient.monitoring) {
       throw Error('Cannot update patient monitoring with "undefined"')
     }
@@ -164,9 +165,9 @@ export default function usePatientProviderCustomHook(): PatientContextResult {
       console.error(error)
       throw Error(`Failed to update patient with id ${patient.userid}`)
     }
-  }, [refresh])
+  }
 
-  const removePatient = useCallback(async (patient: Patient, patientTeam: PatientTeam) => {
+  const removePatient = async (patient: Patient, patientTeam: PatientTeam): Promise<void> => {
     if (patientTeam.status === UserInvitationStatus.pending) {
       const invitation = getInvitation(patientTeam.teamId)
       await cancelInvitation(invitation.id, undefined, invitation.email)
@@ -183,23 +184,23 @@ export default function usePatientProviderCustomHook(): PatientContextResult {
       await flagPatient(patient.userid)
     }
     refresh()
-  }, [cancelInvitation, flagPatient, getFlagPatients, getInvitation, user.id, refresh])
+  }
 
-  const leaveTeam = useCallback(async (teamId: string) => {
+  const leaveTeam = async (teamId: string): Promise<void> => {
     const loggedInUserAsPatient = getPatientById(user.id)
     await PatientApi.removePatient(teamId, loggedInUserAsPatient.userid)
     metrics.send('team_management', 'leave_team')
     refresh()
     refreshTeams()
-  }, [getPatientById, refresh, refreshTeams, user.id])
+  }
 
-  const setPatientMedicalData = useCallback((userId: string, medicalData: MedicalData | null) => {
+  const setPatientMedicalData = (userId: string, medicalData: MedicalData | null): void => {
     const patient = getPatientById(userId)
     if (patient !== null) {
       patient.metadata.medicalData = medicalData
       updatePatient(patient)
     }
-  }, [getPatientById, updatePatient])
+  }
 
   useEffect(() => {
     if (!initialized && user) {
@@ -207,10 +208,10 @@ export default function usePatientProviderCustomHook(): PatientContextResult {
     }
   }, [fetchPatients, initialized, user])
 
-  return useMemo(() => ({
+  return {
     patients: patientList,
     patientsFilterStats,
-    errorMessage,
+    pendingPatientsCount,
     initialized,
     refreshInProgress,
     getPatientByEmail,
@@ -224,5 +225,5 @@ export default function usePatientProviderCustomHook(): PatientContextResult {
     leaveTeam,
     setPatientMedicalData,
     refresh
-  }), [patientList, patientsFilterStats, errorMessage, initialized, refreshInProgress, getPatientByEmail, getPatientById, filterPatients, invitePatient, editPatientRemoteMonitoring, markPatientMessagesAsRead, updatePatientMonitoring, removePatient, leaveTeam, setPatientMedicalData, refresh])
+  }
 }